drf-password-reset/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-filter==23.5
django-sendgrid-v5==1.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sendgrid==6.11.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
uritemplate==4.1.1
